--- src/main/java/Duke.java
@@ -1,21 +1,0 @@
-public class Duke {
-    public static void main(String[] args) {
-        String logo =  " _   _  ___    _    ____\n" +
-                "| \\ | |/ _ \\  / \\  / ___|\n" +
-                "|  \\| | | | |/ _ \\| |\n" +
-                "| |\\  | |_| / ___ \\ |___\n" +
-                "|_| \\_|\\___/_/   \\_\\____|\n";
-        System.out.println("Hello from\n" + logo);
-
-
-        String welcomeMessage = "____________________________________________________________\n" +
-                " Hello! I'm NOAC\n" +
-                " What can I do for you?\n" +
-                "____________________________________________________________\n" +
-                " Bye. Hope to see you again soon!\n" +
-                "____________________________________________________________";
-
-        System.out.println(welcomeMessage);
-
-    }
-}
